increase tests, add docs, fmt

+++ src/lib/BlobParser.sol
@@ -2,12 +2,14 @@
 pragma solidity 0.8.10;
 
 library BlobParser {
-  function bytesToUint256(bytes memory bs, uint start) internal pure returns (uint256) {
-    require(bs.length >= start + 32, "Slicing out of range");
+    function bytesToUint256(bytes memory bs, uint256 start) internal pure returns (uint256) {
+        require(bs.length >= start + 32, "Slicing out of range");
 
-    uint256 x;
-    assembly { x := mload(add(bs, add(0x20, start))) }
+        uint256 x;
+        assembly {
+            x := mload(add(bs, add(0x20, start)))
+        }
 
-    return x;
-  }
+        return x;
+    }
 }
